tid_portal/requirements.txt
djangorestframework
loguru
python-slugify